requirements.txt
altair==5.5.0
altgraph==0.17.4
attrs==25.1.0
babel==2.17.0
blinker==1.9.0
cachetools==5.5.1
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
contourpy==1.3.1
cycler==0.12.1
et_xmlfile==2.0.0
fonttools==4.56.0
gitdb==4.0.12
GitPython==3.1.44
idna==3.10
Jinja2==3.1.5
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
kiwisolver==1.4.8
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.10.0
mdurl==0.1.2
narwhals==1.26.0
numpy==2.2.3
openpyxl==3.1.5
packaging==24.2
pandas==2.2.3
pefile==2023.2.7
pillow==11.1.0
protobuf==5.29.3
pyarrow==19.0.0
pydeck==0.9.1
Pygments==2.19.1
pyinstaller==6.12.0
pyinstaller-hooks-contrib==2025.1
pyotp==2.9.0
pyparsing==3.2.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2025.1
pywin32-ctypes==0.2.3
referencing==0.36.2
requests==2.32.3
rich==13.9.4
rpds-py==0.22.3
setuptools==75.8.0
six==1.17.0
smmap==5.0.2
streamlit==1.42.0
tenacity==9.0.0
tkcalendar==1.6.1
toml==0.10.2
tornado==6.4.2
typing_extensions==4.12.2
tzdata==2025.1
urllib3==2.3.0
watchdog==6.0.0
XlsxWriter==3.2.2
